Валидации
Успешно валидиране на данните за нов потребител и нова стока!
:smiling_imp:

+++ warehouse/src/BUSINESS/exceptions/InvalidPriceException.java
@@ -1,7 +1,7 @@
 package BUSINESS.exceptions;
 
 public class InvalidPriceException extends Exception {
-    public InvalidPriceException(double min, double max) {
+    public InvalidPriceException(int min, int max) {
         super("Цената трябва да е между " + min + " и " + max + "!");
     }
 }
